icore-agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "icore-agent"
version = "0.1.0"
description = "iCore Agent Platform — Strands Agents + mini-SWE-agent"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
dependencies = [
    # Strands Agents SDK (open-source, model-agnostic)
    "strands-agents>=0.1.0",
    "strands-agents-tools>=0.4.0",
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sse-starlette>=2.1.0",
    # Pydantic
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    # LLM routing (mini-SWE-agent model adapter)
    "litellm>=1.50.0",
    # HTTP client
    "httpx>=0.27.0",
    # Env & logging
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    # Redis (conversation memory)
    "redis>=5.0.0",
    "hiredis>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",       # TestClient
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
icore-agent = "icore_agent.main:start"

[tool.hatch.build.targets.wheel]
packages = ["src/icore_agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
